Match editorPane.background color with editor.background
Currently, the editor pane's background is blue which contrasts
with the greenish-gray editor background. This becomes evident
when using the centered editor layout, as well as when viewing
Jupyter Notebooks (see https://github.com/lucafalasco/matchalk/issues/3).

This commit makes the editor pane background color match that of
the editor.